Fix: Properly terminate, fix variable

+++ docs/scripts/playwright_tests.sh
@@ -18,24 +18,51 @@ cd ../..
 
 isdubad=$!
 
+echo "isdubad is running:"
+echo $isdubad
+
 cd client
 
 npm run dev &
 
 client=$!
 
+echo "the client is running:"
+echo $client
+
 sudo /opt/keycloak/bin/kc.sh start-dev --health-enabled=true &
 
 # wait for keycloak to start
 echo "Waiting for keycloak to start..."
-until curl --silent --head -fsS http://localhost:$live/health/ready
+until curl --silent --head -fsS http://localhost:9000/health/ready
 do
   sleep 1
 done
 keycloak=$!
 
+echo "keycloak is running:"
+echo $keycloak
+
 npx playwright test
 
-kill $isdubad
-kill $client
-kill $keycloak
+
+echo "shutting down client, server and keycloak..."
+if [ ! -z "$isdubad" ]; then
+  if ps -p $isdubad; then
+    kill $isdubad
+  fi
+fi
+
+if [ ! -z "$client" ]; then
+  if ps -p $client; then
+    kill $client
+  fi
+fi
+
+if [ ! -z "$keycloak" ]; then
+  if ps -p $keycloak; then
+    kill $keycloak
+  fi
+fi
+
+echo "Tests done"
